attack collision edit
I used a polygon collider to mimic the collision a sword sweep in link to the past would have. I also added a circle collider as a trigger for later object pickup functionality/maybe dialogue prompts (once/if I turn this into a full game instead of a demo). There are a few kinks to work out but it's going moderately well given the time school has been taking up.