collider kink fixes and stop script
The sword collider had some weird effects on movement. It's been fixed pretty after I added an if statement to stop rotation and movement while the sword is out. This was done to emulate other top down games like a Link to The Past. There's also script that prevents the the sword's coroutine from starting up again until the first has finished. I'm not entirely sure why I needed that since the whole point of a coroutine is to not stop until it's done but unity will be unity I suppose.